% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_molly.R
\name{theme_molly}
\alias{theme_molly}
\title{Molly theme package, inspired by the best doggy to have ever existed}
\usage{
theme_molly()
}
\description{
Molly theme package, inspired by the best doggy to have ever existed
}
\examples{
theme_molly()
}
